refactor(cli): extract modal/confirm machinery into CLIModalConfirmMixin (cli.py god-file slice R3)
